+++ package.json
@@ -59,8 +59,6 @@
     "neo4j-graphql-js": "~2.3.1",
     "node-fetch": "~2.3.0",
     "npm-run-all": "~4.1.5",
-    "passport": "~0.4.0",
-    "passport-jwt": "~4.0.0",
     "sanitize-html": "~1.20.0",
     "slug": "~1.0.0",
     "trunc-html": "~1.1.2",
