+++ src/jwt/encode.js
@@ -3,7 +3,7 @@ import jwt from 'jsonwebtoken'
 import ms from 'ms'
 
 // Generate an Access Token for the given User ID
-export default function generateJwt (user) {
+export default function encode (user) {
   const token = jwt.sign(user, process.env.JWT_SECRET, {
     expiresIn: ms('1d'),
     issuer: process.env.GRAPHQL_URI,
